requirements/api.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
mlflow-skinny==2.14.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
httpx==0.27.0
